Fix major GPU coding bug for CFAR CA
In this commit, I have fixed a major bug with respect to GPU coding.
Whenever, you need a scratchpad variable(matrix) for intermediate computes on
the GPU kernal, always pass it as a global memory access such that each thread
has access to its own row/column of the scratchpad array. This is not
required if the intermediate variable is a 1x1 variable. If you pass a
common scratchpad memory to all the threads, there might be conflicts
since all threads have to access the memory simultaneously and this
might result in faulty results at times.
Also, preferably, allot each row of a 2D scratchpad variable to individual thre d instead of columns, since this will allow contiguous memory access for each
thread.

The following are the changes made to the scripts after taking above
into consideration.
1. In the CFAR_lib file, I have made sure that the datatypes of the
variables for the CPU and the GPU version are the same so that the CPU
and GPU outputs match exactly.
2. In the numbaCudaKernels.py file, I'm now passing the scratchpad
variable as a 2D array.

+++ cfar_lib.py
@@ -67,22 +67,22 @@ def CFAR_CA(signal, num_gaurd, num_train, rate_fa):
     signal_shape = len(signal)
     signal_ext = np.hstack((np.flipud(signal[1::]), signal, np.flipud(signal[0:-1])))  # Typically done to handle boundary condition for 1st sample of the Range FFT sample
     #num_gaurd = 5;
-    GuardBandVector = np.zeros(num_gaurd)
+    GuardBandVector = np.zeros(num_gaurd,dtype=np.float32)
     CFAR_Half_Window_Length = num_train + num_gaurd
-    Vector_Ones = np.ones((num_train))
-    CFAR_Window = np.hstack((Vector_Ones,GuardBandVector,np.array([1]),GuardBandVector,Vector_Ones)) 
-    Threshold_Beta = 2*num_train*(rate_fa**(-1/(2*num_train)) -1) # multiplication by a factor of 2  to include valid samples both sides of the CUT
+    Vector_Ones = np.ones((num_train),dtype=np.float32)
+    CFAR_Window = np.hstack((Vector_Ones,GuardBandVector,np.array([1],dtype=np.float32),GuardBandVector,Vector_Ones)) 
+    Threshold_Beta = np.float32(2*num_train*(rate_fa**(-1/(2*num_train)) -1)) # multiplication by a factor of 2  to include valid samples both sides of the CUT
     count = 0
-    Target_BoolVector = np.zeros((signal_shape)).astype('int')
+    Target_BoolVector = np.zeros((signal_shape)).astype('int32')
     for SignalIndex in np.arange(signal_shape-1,2*signal_shape-1):
-      CUT = signal_ext[SignalIndex]
-      if CUT >= np.amax(np.array([signal_ext[SignalIndex+1], signal_ext[SignalIndex-1]])):
-          Temp = CFAR_Window*signal_ext[SignalIndex-CFAR_Half_Window_Length:SignalIndex+CFAR_Half_Window_Length+1]
-          Temp_CUT_removed = np.hstack((Temp[0:CFAR_Half_Window_Length],Temp[CFAR_Half_Window_Length+1::]))
-          noise_power = np.sum(Temp_CUT_removed)/(2*num_train)
-          if (CUT > Threshold_Beta*noise_power):
-            Target_BoolVector[count] = 1
-      count = count + 1
+        CUT = signal_ext[SignalIndex]
+        if CUT >= np.amax(np.array([signal_ext[SignalIndex+1], signal_ext[SignalIndex-1]])):
+            Temp = CFAR_Window*signal_ext[SignalIndex-CFAR_Half_Window_Length:SignalIndex+CFAR_Half_Window_Length+1]
+            Temp_CUT_removed = np.hstack((Temp[0:CFAR_Half_Window_Length],Temp[CFAR_Half_Window_Length+1::]))
+            noise_power = np.sum(Temp_CUT_removed)/(2*num_train)
+            if (CUT > Threshold_Beta*noise_power):
+              Target_BoolVector[count] = 1
+        count = count + 1
 
 
     return Target_BoolVector
